ProyectoResidencia/requirements.txt
openpyxl
python-dotenv
requests
pandas
flask